Rename package to satisfy Checkstyle
