+++ src/main/java/at/ac/tuwien/kr/alpha/solver/heuristics/activity/DefaultBodyActivityProvider.java
@@ -23,7 +23,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package at.ac.tuwien.kr.alpha.solver.heuristics.body_activity;
+package at.ac.tuwien.kr.alpha.solver.heuristics.activity;
 
 import org.apache.commons.collections4.MultiValuedMap;
